Add CMake build system
meson sounds cool, but I've spent too long trying to get it to build
with EGL (on Windows).  CMake on the other hand works fine.

I've only tested this on Windows.  Shouldn't be too far off for unix
platforms.

There are some things missing here that are present in the meson build,
like which warnings to ignore.